+++ subscription/util/blocking/catchup-subscription/src/main/java/org/occurrent/subscription/util/blocking/catchup/subscription/CatchupSupportingBlockingSubscriptionConfig.java
@@ -19,6 +19,7 @@ package org.occurrent.subscription.util.blocking.catchup.subscription;
 import io.cloudevents.CloudEvent;
 import org.occurrent.eventstore.api.blocking.EventStoreQueries;
 import org.occurrent.subscription.api.blocking.Subscription;
+import org.occurrent.subscription.util.predicate.EveryN;
 
 import java.util.Objects;
 import java.util.function.Predicate;
@@ -48,12 +49,12 @@ public class CatchupSupportingBlockingSubscriptionConfig {
 
 
     /**
-     * @param cacheSize                            The number of cloud events id's to persistedPositionForEveryNCloudEvent in-memory when switching from "catch-up" mode (i.e. querying the {@link EventStoreQueries} API)
-     *                                             and "subscription" mode ({@link Subscription}). The cache is needed to reduce the number of duplicate events the occurs when switching.
-     * @param persistedPositionForEveryNCloudEvent Persist the position of every N cloud event so that it's possible to avoid restarting from scratch when subscription is restarted.
+     * @param cacheSize                          The number of cloud events id's to store in-memory when switching from "catch-up" mode (i.e. querying the {@link EventStoreQueries} API)
+     *                                           and "subscription" mode ({@link Subscription}). The cache is needed to reduce the number of duplicate events the occurs when switching.
+     * @param persistPositionForEveryNCloudEvent Persist the position of every N cloud event so that it's possible to avoid restarting from scratch when subscription is restarted.
      */
-    public CatchupSupportingBlockingSubscriptionConfig(int cacheSize, int persistedPositionForEveryNCloudEvent) {
-        this(cacheSize, new EveryN(persistedPositionForEveryNCloudEvent));
+    public CatchupSupportingBlockingSubscriptionConfig(int cacheSize, int persistPositionForEveryNCloudEvent) {
+        this(cacheSize, new EveryN(persistPositionForEveryNCloudEvent));
     }